the number of streams is no longer limited (Jira SN-14) (well actually it _is_ limited, but at 13119 streams *g*)
git-svn-id: http://red5.googlecode.com/svn/java/server/trunk@1710 1b6495e4-3631-0410-8e05-8f51eee8b9cc

+++ src/org/red5/server/net/rtmp/BaseRTMPHandler.java
@@ -198,7 +198,7 @@ public abstract class BaseRTMPHandler implements IRTMPHandler, Constants, Status
 		conn.messageSent((Packet) message);
 
 		Packet sent = (Packet) message;
-		final byte channelId = sent.getHeader().getChannelId();
+		final int channelId = sent.getHeader().getChannelId();
 		final IClientStream stream = conn.getStreamByChannelId(channelId);
 		// XXX we'd better use new event model for notification
 		if (stream != null && (stream instanceof PlaylistSubscriberStream)) {
